refactor(config): split agent sources so embedders link only the source types they use
- config.Load is now YAML-only. HCL support becomes a source decorator,
  hcl.NewSource, applied by the resolver; embedders wrap sources that may
  carry HCL themselves.
- The OCI source moves to pkg/config/ocisource (ocisource.New).
- Resolve, ResolveSources, ResolveAlias and BuiltinAgentNames move to
  pkg/config/sources together with the embedded built-in agents. This is
  the package that knows every source type.
- teamloader no longer hardwires the resolver for external sub-agents:
  WithSourceResolver supplies it, and loaderdefaults.Opts wires
  sources.Resolve so CLI behaviour is unchanged.

pkg/config drops hcl, go-cty, levenshtein, textseg, wordwrap,
klauspost/compress and image-spec, and only reaches
go-containerregistry/pkg/name for reference parsing.
