+++ cmd/root/backend.go
@@ -8,6 +8,7 @@ import (
 	"sync"
 
 	"github.com/docker/docker-agent/pkg/config"
+	"github.com/docker/docker-agent/pkg/config/sources"
 	pathx "github.com/docker/docker-agent/pkg/path"
 	"github.com/docker/docker-agent/pkg/runtime"
 	"github.com/docker/docker-agent/pkg/session"
@@ -51,7 +52,7 @@ func (f *runExecFlags) selectBackend(agentFileName string) (backend, error) {
 	if f.remoteAddress != "" {
 		return &remoteBackend{flags: f, agentFileName: agentFileName}, nil
 	}
-	agentSource, err := config.Resolve(agentFileName, f.runConfig.EnvProvider())
+	agentSource, err := sources.Resolve(agentFileName, f.runConfig.EnvProvider())
 	if err != nil {
 		return nil, err
 	}
