+++ pkg/a2a/adapter_test.go
@@ -8,6 +8,7 @@ import (
 	"google.golang.org/genai"
 
 	"github.com/docker/docker-agent/pkg/config"
+	"github.com/docker/docker-agent/pkg/config/sources"
 	"github.com/docker/docker-agent/pkg/servesafety"
 	"github.com/docker/docker-agent/pkg/session"
 	"github.com/docker/docker-agent/pkg/teamloader"
@@ -17,7 +18,7 @@ import (
 func TestNewDockerAgentAdapter(t *testing.T) {
 	t.Setenv("OPENAI_API_KEY", "DUMMY")
 
-	agentSource, err := config.Resolve("testdata/basic.yaml", nil)
+	agentSource, err := sources.Resolve("testdata/basic.yaml", nil)
 	require.NoError(t, err)
 
 	team, err := teamloader.Load(t.Context(), agentSource, &config.RuntimeConfig{}, loaderdefaults.Opts()...)
@@ -36,7 +37,7 @@ func TestNewDockerAgentAdapter(t *testing.T) {
 func TestNewCAgentAdapter_NonExistent(t *testing.T) {
 	t.Setenv("OPENAI_API_KEY", "DUMMY")
 
-	agentSource, err := config.Resolve("testdata/basic.yaml", nil)
+	agentSource, err := sources.Resolve("testdata/basic.yaml", nil)
 	require.NoError(t, err)
 
 	team, err := teamloader.Load(t.Context(), agentSource, &config.RuntimeConfig{}, loaderdefaults.Opts()...)
