+++ pkg/a2a/server.go
@@ -15,6 +15,7 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
+
 	// go.mod pins adk/v2 to pseudo-version v2.2.1-0.20260818092052-16e33673bc9e:
 	// OTel Logs 0.21 support (google/adk-go#1335) is merged upstream but not yet
 	// tagged. Replace with the first v2 tag that contains that commit.
@@ -23,6 +24,7 @@ import (
 	adksession "google.golang.org/adk/v2/session"
 
 	"github.com/docker/docker-agent/pkg/config"
+	"github.com/docker/docker-agent/pkg/config/sources"
 	"github.com/docker/docker-agent/pkg/httpsec"
 	pathx "github.com/docker/docker-agent/pkg/path"
 	"github.com/docker/docker-agent/pkg/servesafety"
@@ -57,7 +59,7 @@ type RunOptions struct {
 func Run(ctx context.Context, agentFilename, agentName, sessionDB string, runConfig *config.RuntimeConfig, ln net.Listener, options RunOptions) error {
 	slog.DebugContext(ctx, "Starting A2A server", "source", agentFilename, "agent", agentName, "addr", ln.Addr().String())
 
-	agentSource, err := config.Resolve(agentFilename, nil)
+	agentSource, err := sources.Resolve(agentFilename, nil)
 	if err != nil {
 		return err
 	}
