+++ pkg/config/sources/builtin-agents/coder.yaml
