+++ pkg/server/mcp_oauth_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/docker/docker-agent/pkg/config"
+	"github.com/docker/docker-agent/pkg/config/sources"
 )
 
 // httpDoStatus is a slim variant of httpDo that exposes the response
@@ -42,7 +43,7 @@ func startServerBare(t *testing.T, ctx context.Context) string {
 	t.Helper()
 	var store mockStore
 	runConfig := config.RuntimeConfig{}
-	sources, err := config.ResolveSources(t.TempDir(), nil)
+	sources, err := sources.ResolveSources(t.TempDir(), nil)
 	require.NoError(t, err)
 	srv, err := New(ctx, store, &runConfig, 0, sources, "", 0)
 	require.NoError(t, err)
