+++ pkg/server/server_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/docker/docker-agent/pkg/api"
 	"github.com/docker/docker-agent/pkg/chat"
 	"github.com/docker/docker-agent/pkg/config"
+	"github.com/docker/docker-agent/pkg/config/sources"
 	"github.com/docker/docker-agent/pkg/session"
 )
 
@@ -204,7 +205,7 @@ func startServer(t *testing.T, ctx context.Context, agentsDir string) string {
 	var store mockStore
 	runConfig := config.RuntimeConfig{}
 
-	sources, err := config.ResolveSources(agentsDir, nil)
+	sources, err := sources.ResolveSources(agentsDir, nil)
 	require.NoError(t, err)
 	srv, err := New(ctx, store, &runConfig, 0, sources, "", 0)
 	require.NoError(t, err)
@@ -541,7 +542,7 @@ func startServerWithStore(t *testing.T, ctx context.Context, agentsDir string, s
 
 	runConfig := config.RuntimeConfig{}
 
-	sources, err := config.ResolveSources(agentsDir, nil)
+	sources, err := sources.ResolveSources(agentsDir, nil)
 	require.NoError(t, err)
 	srv, err := New(ctx, store, &runConfig, 0, sources, "", 0)
 	require.NoError(t, err)
